Finish my existence(¬_¬ )

+++ src/main/java/com/para/repo/GameRepo.java
@@ -4,8 +4,7 @@ import com.para.games.Game;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
 public interface GameRepo extends JpaRepository<Game, Integer> {
-    void deleteGameById(Long id);
-    Optional<Game> findGameById(Long id);
+    void deleteGameById(Integer id);
+    Optional<Game> findGameById(Integer id);
 }
